+++ src/Odometer/Odometer.jsx
@@ -1,0 +1,14 @@
+import React, { PureComponent } from 'react';
+import PropTypes from 'prop-types';
+
+/**
+ * An odometer that would display a positive numer ina rolling fashion
+ */
+export default class Odometer extends PureComponent {
+  render () {
+    return (
+      <svg>
+      </svg>
+    );
+  }
+}