+++ src/ProgressBar/ProgressBar.test.js
@@ -13,6 +13,6 @@ describe('ProgressBar', () => {
 
   it('should changes value on props', () => {
     bar.setProps({ value: .25 });
-    expect(bar.find('.status-progress-bar').prop('style').transform).toMatch(/translateX\(-75%\)/);
+    expect(bar.find('.status-progress-bar-value').prop('style').transform).toMatch(/translateX\(-75%\)/);
   });
 });