+++ CHANGELOG.md
@@ -23,6 +23,7 @@ The format is based on [Keep a Changelog](http://keepachangelog.com/en/1.0.0/) a
 ### Changed
 * add `title` prop to ProgressBar
 * updated packages
+* Updated ProgressBar to accept `value` prop to be an array of values
 
 ## [1.0.1] - 2018-03-01
 ### Changed